requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
pandas
numpy
